+++ src/main/java/org/ktb/matajo/service/chat/ChatMessageServiceImpl.java
@@ -1,11 +1,12 @@
 package org.ktb.matajo.service.chat;
 
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.ktb.matajo.dto.chat.ChatMessageRequestDto;
 import org.ktb.matajo.dto.chat.ChatMessageResponseDto;
 import org.ktb.matajo.entity.*;
+import org.ktb.matajo.global.error.code.ErrorCode;
+import org.ktb.matajo.global.error.exception.BusinessException;
 import org.ktb.matajo.repository.ChatMessageRepository;
 import org.ktb.matajo.repository.ChatRoomRepository;
 import org.ktb.matajo.repository.UserRepository;
@@ -37,16 +38,32 @@ public class ChatMessageServiceImpl implements ChatMessageService {
     @Transactional
     public ChatMessageResponseDto saveMessage(Long roomId, ChatMessageRequestDto messageDto) {
 
-        log.info("메시지 요청 DTO: {}", messageDto);
-        log.info("senderId: {}, 타입: {}", messageDto.getSenderId(), messageDto.getSenderId() != null ? messageDto.getSenderId().getClass().getName() : "null");
+        validateRoomId(roomId);
+
+        // DTO 추가 메서드를 활용한 검증
+        if(messageDto.isImageTypeWithEmptyContent()) {
+            log.error("이미지 타입 메시지의 내용이 비어있습니다");
+            throw new BusinessException(ErrorCode.INVALID_IMAGE_CONTENT);
+        }
+
+        if(!messageDto.isValidImageUrl()) {
+            log.error("유효하지 않은 이미지 URL 형식: {}", messageDto.getContent());
+            throw new BusinessException(ErrorCode.INVALID_IMAGE_URL);
+        }
 
         // 채팅방 조회
         ChatRoom chatRoom = chatRoomRepository.findById(roomId)
-                .orElseThrow(() -> new EntityNotFoundException("채팅방을 찾을 수 없습니다. ID: " + roomId));
+                .orElseThrow(() -> {
+                    log.error("채팅방을 찾을 수 없습니다");
+                    return new BusinessException(ErrorCode.CHAT_ROOM_NOT_FOUND);
+                });
 
         // 사용자 조회
         User sender = userRepository.findById(messageDto.getSenderId())
-                .orElseThrow(() -> new EntityNotFoundException("사용자를 찾을 수 없습니다. ID: " + messageDto.getSenderId()));
+                .orElseThrow(() -> {
+                    log.error("사용자를 찾을 수 없습니다");
+                    return new BusinessException(ErrorCode.USER_NOT_FOUND);
+                });
 
         // 채팅 메시지 생성
         ChatMessage chatMessage = ChatMessage.builder()
@@ -64,8 +81,12 @@ public class ChatMessageServiceImpl implements ChatMessageService {
         // 응답 DTO 생성
         ChatMessageResponseDto responseDto = convertToChatMessageResponseDto(savedMessage);
 
-        // Redis에 캐싱
-        redisChatMessageService.cacheMessage(roomId, responseDto);
+        // 안전한 캐싱 (예외가 발생해도 메인 로직에 영향 없게)
+        try {
+            redisChatMessageService.cacheMessage(roomId, responseDto);
+        } catch (Exception e) {
+            log.warn("메시지 캐싱 실패 (무시됨): {}", e.getMessage());
+        }
 
         return responseDto;
     }
@@ -75,6 +96,13 @@ public class ChatMessageServiceImpl implements ChatMessageService {
      */
     @Override
     public List<ChatMessageResponseDto> getChatMessages(Long roomId, int page, int size) {
+
+        validateRoomId(roomId);
+
+        if (page < 0 || size <= 0 || size > 100) {
+            log.error("페이지네이션 파라미터가 유효하지 않습니다: page={}, size={}", page, size);
+            throw new BusinessException(ErrorCode.INVALID_OFFSET_OR_LIMIT);
+        }
         // 첫 페이지이면 캐시에서 먼저 조회 시도
         if (page == 0) {
             List<ChatMessageResponseDto> cachedMessages = redisChatMessageService.getCachedMessages(roomId, size);
@@ -96,9 +124,13 @@ public class ChatMessageServiceImpl implements ChatMessageService {
                 .map(this::convertToChatMessageResponseDto)
                 .collect(Collectors.toList());
 
-        // 첫 페이지 결과를 Redis에 캐싱
+        // 첫 페이지 결과를 Redis에 캐싱 - 예외 처리 추가
         if (page == 0 && !messageDtos.isEmpty()) {
-            redisChatMessageService.cacheMessages(roomId, messageDtos);
+            try {
+                redisChatMessageService.cacheMessages(roomId, messageDtos);
+            } catch (Exception e) {
+                log.warn("메시지 목록 캐싱 실패 (무시됨): {}", e.getMessage());
+            }
         }
 
         return messageDtos;
@@ -110,6 +142,15 @@ public class ChatMessageServiceImpl implements ChatMessageService {
     @Override
     @Transactional
     public void markMessagesAsRead(Long roomId, Long userId) {
+
+        validateRoomId(roomId);
+
+        // 파라미터 검증
+        if (userId == null || userId <= 0) {
+            log.error("유효하지 않은 userId 값입니다: {}", userId);
+            throw new BusinessException(ErrorCode.INVALID_USER_ID);
+        }
+
         // 읽지 않은 메시지 중 다른 사용자가 보낸 메시지만 읽음 처리
         List<ChatMessage> unreadMessages = chatMessageRepository.findUnreadMessagesForUser(roomId, userId);
 
@@ -119,8 +160,12 @@ public class ChatMessageServiceImpl implements ChatMessageService {
 
         chatMessageRepository.saveAll(unreadMessages);
 
-        // Redis 캐시 업데이트 (캐시 무효화)
-        redisChatMessageService.invalidateCache(roomId);
+        // Redis 캐시 업데이트 (캐시 무효화) - 예외 처리 추가
+        try {
+            redisChatMessageService.invalidateCache(roomId);
+        } catch (Exception e) {
+            log.warn("캐시 무효화 실패 (무시됨): {}", e.getMessage());
+        }
     }
 
     /**
@@ -137,5 +182,18 @@ public class ChatMessageServiceImpl implements ChatMessageService {
                 .readStatus(message.isReadStatus())
                 .createdAt(message.getCreatedAt())
                 .build();
+    }
+
+    private void validateRoomId(Long roomId) {
+        // 간단한 파라미터 검증
+        if (roomId == null) {
+            log.error("roomId가 null입니다");
+            throw new BusinessException(ErrorCode.INVALID_INPUT_VALUE);
+        }
+
+        if (roomId <= 0) {
+            log.error("유효하지 않은 roomId 값입니다: {}", roomId);
+            throw new BusinessException(ErrorCode.INVALID_INPUT_VALUE);
+        }
     }
 }
